+++ code/tests/unit.h
@@ -31,7 +31,7 @@ static int _g_errors = 0;
     _errors += _lasterr; \
     zpl_printf(" * [%s]: It %s %s\n", (_lasterr) ? "\x1B[31mFAIL\x1B[0m" : "\x1B[32mPASS\x1B[0m", desc, _errstr);
 
-#define FAIL(a, b)                             { _errstr = zpl_bprintf("\n\n\tassert: \x1B[31m%s:%d %s %s:%d\x1B[0m\n\tat %s:%d\n", #a, (int)a, (a == b)?"==":"!=", #b, b, __FILE__, __LINE__); _lasterr = 1; break; }
+#define FAIL(a, b)                             { _errstr = zpl_bprintf("\n\n\tassert: \x1B[31m%s:%d %s %s:%d\x1B[0m\n\tat %s:%d\n", #a, (int)((intptr_t)a), (a == b)?"==":"!=", #b, b, __FILE__, __LINE__); _lasterr = 1; break; }
 #define STRFAIL(a, b)                          { _errstr = zpl_bprintf("\n\n\tassert: \x1B[31m%s:%s %s %s:%s\x1B[0m\n\tat %s:%d\n", #a, (char *)a, (!strcmp(a,b))?"==":"!=", #b, b, __FILE__, __LINE__); _lasterr = 1; break; }
 #define EQUALS(a, b)        if (a != b)        { FAIL(a, b); }
 #define STREQUALS(a, b)     if (!!strcmp(a,b)) { STRFAIL(a, b); }
